fixing int to int64

+++ models/prescription.go
@@ -15,8 +15,8 @@ type Prescription struct {
 	ID           int     `json:"id"`
 	TreatmentId  int     `json:"treatment_id"`
 	MedicationId int     `json:"medication_id"`
-	StartingAt   int     `json:"starting_at"`
-	FinishingAt  int     `json:"finishing_at"`
+	StartingAt   int64   `json:"starting_at"`
+	FinishingAt  int64   `json:"finishing_at"`
 	Frequency    int     `json:"frequency"`
 	Receipt      Receipt `json:"receipt"`
 }
